+++ src/main/kotlin/com/pswidersk/sdkimportplugin/JavaSdkProcessor.kt
@@ -1,7 +1,5 @@
 package com.pswidersk.sdkimportplugin
 
-import com.intellij.notification.Notification
-import com.intellij.notification.NotificationType
 import com.intellij.openapi.module.Module
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.projectRoots.JavaSdk
@@ -9,13 +7,12 @@ import com.intellij.openapi.projectRoots.ProjectJdkTable
 import com.intellij.openapi.projectRoots.Sdk
 import com.intellij.openapi.roots.ModuleRootModificationUtil
 import com.intellij.openapi.roots.ProjectRootManager
-import com.pswidersk.sdkimportplugin.SdkImportBundle.message
 
 
 class JavaSdkProcessor : SdkProcessor {
 
     override fun applySdk(project: Project, sdkConfig: SdkImportConfigEntry) {
-        if (sdkConfig.type == SdkType.JAVA) {
+        if (sdkConfig.type == JAVA_SDK_TYPE) {
             addJdk(project, sdkConfig)
         }
     }
@@ -30,7 +27,7 @@ class JavaSdkProcessor : SdkProcessor {
             withWriteAction {
                 sdkTable.addJdk(jdk)
             }
-            notifyAboutNewSdk(project, jdkName)
+            newJavaSdkNotif(project, jdkName)
             jdk
         }
         project.withModule(sdkConfig.module) {
@@ -48,40 +45,14 @@ class JavaSdkProcessor : SdkProcessor {
         if (project.name == sdkConfig.module) {
             withWriteAction {
                 ProjectRootManager.getInstance(project).projectSdk = sdk
-                notifyAboutProjectSdkChange(project, sdk.name)
+                changedProjectJavaSdkNotif(project, sdk.name)
             }
         }
     }
 
     private fun setModuleSdk(module: Module, sdk: Sdk) {
         ModuleRootModificationUtil.setModuleSdk(module, sdk)
-        notifyAboutModuleSdkChange(module, sdk.name)
+        changedModuleJavaSdkNotif(module, sdk.name)
     }
 
-    private fun notifyAboutNewSdk(project: Project, sdkName: String) {
-        Notification(
-            NOTIFICATION_GROUP,
-            message("notification.newJavaSdk.title"),
-            message("notification.newJavaSdk.content", sdkName),
-            NotificationType.INFORMATION
-        ).notify(project)
-    }
-
-    private fun notifyAboutModuleSdkChange(module: Module, sdkName: String) {
-        Notification(
-            NOTIFICATION_GROUP,
-            message("notification.newModuleJavaSdk.title"),
-            message("notification.newModuleJavaSdk.content", module.name, sdkName),
-            NotificationType.INFORMATION
-        ).notify(module.project)
-    }
-
-    private fun notifyAboutProjectSdkChange(project: Project, sdkName: String) {
-        Notification(
-            NOTIFICATION_GROUP,
-            message("notification.newProjectJavaSdk.title"),
-            message("notification.newProjectJavaSdk.content", project.name, sdkName),
-            NotificationType.INFORMATION
-        ).notify(project)
-    }
 }