+++ src/test/kotlin/com/pswidersk/sdkimportplugin/SdkImportServiceTest.kt
@@ -1,11 +1,12 @@
 package com.pswidersk.sdkimportplugin
 
+import com.intellij.notification.Notification
+import com.intellij.notification.impl.NotificationsManagerImpl
 import com.intellij.openapi.application.runWriteAction
 import com.intellij.openapi.components.service
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.projectRoots.ProjectJdkTable
 import com.intellij.openapi.vfs.newvfs.impl.VfsRootAccess
-import com.intellij.openapi.vfs.writeText
 import com.intellij.testFramework.junit5.RunInEdt
 import com.intellij.testFramework.junit5.TestApplication
 import com.intellij.testFramework.rules.ProjectModelExtension
@@ -18,7 +19,6 @@ import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.extension.RegisterExtension
 import java.io.File
 
-private const val SDK_IMPORT_REF = ".idea/sdk-import.yml"
 private const val TEST_MODULE_NAME = "sample-python-module"
 
 @TestApplication
@@ -28,6 +28,12 @@ class SdkImportServiceTest {
     @JvmField
     @RegisterExtension
     val projectModel: ProjectModelExtension = ProjectModelExtension()
+
+    private val ideaDir: File
+        get() = projectModel.baseProjectDir.root.resolve(".idea").also { it.mkdir() }
+
+    private val sdkImportFile: File
+        get() = ideaDir.resolve("sdk-import.yml")
 
     private val project: Project
         get() = projectModel.project
@@ -41,12 +47,12 @@ class SdkImportServiceTest {
     @BeforeEach
     fun setUp() {
         projectModel.createModule(TEST_MODULE_NAME)
-        mockPythonSdk()
     }
 
     @Test
-    fun `new SDK is imported`() {
+    fun `new Python SDK is imported`() {
         // given
+        mockPythonSdk()
         val projectService = project.service<SdkImportService>()
 
         // when
@@ -58,13 +64,46 @@ class SdkImportServiceTest {
         assertThat(rootModule.pythonSdk?.name).isEqualTo("Python env: $pythonPath")
     }
 
+    @Test
+    fun `error notification if displayed`() {
+        // given
+        val projectService = project.service<SdkImportService>()
+        val notificationsManager = NotificationsManagerImpl.getNotificationsManager()
+        saveErroneousConfigFile()
+
+        // when
+        projectService.runImport()
+
+        // then
+        val notifications = notificationsManager.getNotificationsOfType(Notification::class.java, project)
+        assertThat(notifications).hasSize(1)
+        with(notifications.first()) {
+            assertThat(title).isEqualTo("SDK-Import -> an exception occurred.")
+            assertThat(content).startsWith("Exception message: `Cannot create property=import")
+        }
+    }
+
     @AfterEach
     fun tearDown() {
         clearSdks()
+        sdkImportFile.writeText("")
+    }
+
+    private fun saveErroneousConfigFile() {
+        runWriteAction {
+            sdkImportFile.writeText(
+                """
+                import:
+                  - type: NON_SUPPORTED
+                    non-parsable: someValue
+                    module: $TEST_MODULE_NAME
+                """.trimIndent()
+            )
+        }
     }
 
     private fun findPythonPath(buildProjectDir: String): String {
-        val testSdkImportConfigFile = File(buildProjectDir).resolve(SDK_IMPORT_REF)
+        val testSdkImportConfigFile = File(buildProjectDir).resolve(".idea/sdk-import.yml")
         val testSdkImportConfig = testSdkImportConfigFile.loadAsYamlImportConfig()
         return testSdkImportConfig.import.first().path
     }
@@ -72,12 +111,12 @@ class SdkImportServiceTest {
     private fun mockPythonSdk() {
         VfsRootAccess.allowRootAccess(project, buildProjectDir)
         runWriteAction {
-            projectModel.baseProjectDir.newVirtualFile(SDK_IMPORT_REF).writeText(
+            sdkImportFile.writeText(
                 """
-    import:
-      - type: PYTHON
-        path: $pythonPath
-        module: $TEST_MODULE_NAME
+                import:
+                  - type: PYTHON
+                    path: $pythonPath
+                    module: $TEST_MODULE_NAME
                 """.trimIndent()
             )
         }