+++ CHANGELOG.md
@@ -4,6 +4,14 @@
 
 ## [Unreleased]
 
+### Added
+
+- Gradle Python new project generation wizard.
+
+### Changed
+
+- Config loading errors notifications handling.
+
 ## [1.0.1] - 2024-04-27
 
 ### Added
@@ -12,4 +20,5 @@
 - Using Snake YAML instead of Jackson for the sake of backward compatibility.
 
 [Unreleased]: https://github.com/PrzemyslawSwiderski/sdk-import-plugin/compare/v1.0.1...HEAD
+
 [1.0.1]: https://github.com/PrzemyslawSwiderski/sdk-import-plugin/commits/v1.0.1